section_13-pierrot/springsecsection_13-pierrot/src/main/java/com/eazybytes/controller/CardsController.java
package com.eazybytes.controller;

import com.eazybytes.model.Cards;
import com.eazybytes.model.Customer;
import com.eazybytes.repository.CardsRepository;
import com.eazybytes.repository.CustomerRepository;
import java.util.Collections;
import org.springframework.util.CollectionUtils;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;

import java.util.List;

@RestController
public class CardsController {

    private final CardsRepository cardsRepository;

    private final CustomerRepository customerRepository;

    public CardsController(CardsRepository cardsRepository, CustomerRepository customerRepository) {
        this.cardsRepository = cardsRepository;
        this.customerRepository = customerRepository;
    }

    @GetMapping("/myCards")
    public List<Cards> getCardDetails(@RequestParam String email) {

        List<Customer> custByEmail = customerRepository.findByEmail(email);
        if (!CollectionUtils.isEmpty(custByEmail)) {
            return cardsRepository.findByCustomerId(custByEmail.get(0).getId());
        }

        return Collections.emptyList();
    }

}
